separated overlay file, created config vars for info overlay
