+++ Common/Settings.ini
@@ -12,6 +12,9 @@ WndModeBorder = 1
 ; Detects every resolution your display offers and makes them selectable choices (along with showing their aspect ratios) in the game's Options > Advanced Options menu. 
 DynamicResolution = 1 
 
+; Shows an overlay containing game stats, with number of kills, items etc.
+EnableInfoOverlay = 1
+
 [Fullscreen Media]
 ; Adds compatibility for the project's Image Enhancement Pack to have its visuals display correctly. Must be enabled if using the Image Enhancement Pack. The scaling works for both the original and enhanced fullscreen images. Auto: Automatically chooses "fit" or "fill" depending on the aspect ratio for the best presentation. Fit: Displays the fullscreen images without any cropping. Letterboxes or pillarboxes may appear depending on the aspect ratio. Fill: Scales the fullscreen images to fill the display area, removing any letterboxes or pillarboxes, up to a 16:9 aspect ratio. 
 FullscreenImages = 3
