+++ Wrappers/d3d8/Overlay.h
@@ -1,0 +1,40 @@
+#pragma once
+#include "d3d8wrapper.h"
+#include <sstream>
+
+class Overlay
+{
+public:
+	struct D3D8TEXT
+	{
+		LPCSTR   String;
+		RECT       Rect;
+		DWORD    Format;
+	};
+
+	void DrawDebugOverlay(LPDIRECT3DDEVICE8 ProxyInterface);
+	void ResetFont();
+
+private:
+	std::string ActionDifficulty[4]
+	{
+		"Beginner",
+		"Easy",
+		"Normal",
+		"Hard"
+	};
+
+	std::string RiddleDifficulty[4]
+	{
+		"Easy",
+		"Normal",
+		"Hard",
+		"Extra"
+	};
+
+	void DrawDebugText(LPDIRECT3DDEVICE8 ProxyInterface, D3D8TEXT TextStruct);
+	std::string IntToHexStr(int IntValue);
+	std::string FloatToStr(float FloatValue, int precision);
+	
+};
+
